refactored maps.py: moved expression related functions and variable from gates.py to expressions.py
