+++ src/pyqasm/maps/gates.py
@@ -19,64 +19,12 @@ Module mapping supported QASM gates to lower level gate operations.
 from typing import Callable, Union
 
 import numpy as np
-from openqasm3.ast import (
-    FloatLiteral,
-    Identifier,
-    IndexedIdentifier,
-    QuantumGate,
-    QuantumPhase,
-)
+from openqasm3.ast import FloatLiteral, Identifier, IndexedIdentifier, QuantumGate, QuantumPhase
 
 from pyqasm.elements import InversionOp
 from pyqasm.exceptions import ValidationError
 from pyqasm.linalg import kak_decomposition_angles
-from pyqasm.maps import OperatorFunction
 from pyqasm.maps.expressions import CONSTANTS_MAP
-
-
-OPERATOR_MAP: dict[str, OperatorFunction] = {
-    "+": lambda x, y: x + y,
-    "-": lambda x, y: x - y,
-    "*": lambda x, y: x * y,
-    "/": lambda x, y: x / y,
-    "%": lambda x, y: x % y,
-    "==": lambda x, y: x == y,
-    "!=": lambda x, y: x != y,
-    "<": lambda x, y: x < y,
-    ">": lambda x, y: x > y,
-    "<=": lambda x, y: x <= y,
-    ">=": lambda x, y: x >= y,
-    "&&": lambda x, y: x and y,
-    "||": lambda x, y: x or y,
-    "^": lambda x, y: x ^ y,
-    "&": lambda x, y: x & y,
-    "|": lambda x, y: x | y,
-    "<<": lambda x, y: x << y,
-    ">>": lambda x, y: x >> y,
-    "~": lambda x: ~x,
-    "!": lambda x: not x,
-    "UMINUS": lambda x: -x,
-}
-
-
-def qasm3_expression_op_map(op_name: str, *args) -> Union[float, int, bool]:
-    """
-    Return the result of applying the given operator to the given operands.
-
-    Args:
-        op_name (str): The operator name.
-        *args: The operands of type Union[int, float, bool]
-                1. For unary operators, a single operand (e.g., ~3)
-                2. For binary operators, two operands (e.g., 3 + 2)
-
-    Returns:
-        (Union[float, int, bool]): The result of applying the operator to the operands.
-    """
-    try:
-        operator = OPERATOR_MAP[op_name]
-        return operator(*args)
-    except KeyError as exc:
-        raise ValidationError(f"Unsupported / undeclared QASM operator: {op_name}") from exc
 
 
 def u3_gate(
